apps/api/pyproject.toml
[project]
name = "drinkx-crm-api"
version = "0.1.0"
description = "DrinkX CRM API — FastAPI backend, package-per-domain"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.1",
    "httpx>=0.28.1",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.19.2",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0",
    "aiosmtplib>=3.0.2",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.150.0",
    "cryptography>=43.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "types-python-jose>=3.3.4.20240106",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "RUF", "ASYNC"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
